Implemented PlayerStatsModal
