+++ assets/js/pages/database/player_stats/index.js
@@ -1,11 +1,13 @@
-import playerStatsModal from './playerStatsModal.js';
+import blackAutocomplete from '../../../elements/blackAutocomplete.js';
 import boardActionsDropdown from '../../../elements/boardActionsDropdown.js';
 import gameStudyDropdown from '../../../elements/gameStudyDropdown.js';
 import historyButtons from '../../../elements/historyButtons.js';
+import playerStatsModal from '../../../elements/playerStatsModal.js';
+import whiteAutocomplete from '../../../elements/whiteAutocomplete.js';
 import ws from '../../../sanWs.js';
 
 await ws.connect();
 
 localStorage.clear();
 
-playerStatsModal.modal.show();
+playerStatsModal.props.modal.show();
